working on updating ft

+++ app/Http/Controllers/Admin/UsersController.php
@@ -102,7 +102,6 @@ class UsersController extends Controller
     public function update(Request $request, $id)
     {
         $userParams = json_decode($request->input('userParams'));
-        $reportCards = $userParams->newReportCards;
         $user = User::find($userParams->user_id);
         // $this->validate($request, array(
         //     'name' => 'required|max:255',
@@ -130,34 +129,50 @@ class UsersController extends Controller
         }
 
         // Update existing report cards
-        foreach($reportCards as $report){
-            if($report->deleted == false){
-                is_numeric($report->id) ? $currentReport = (ReportCard::find($report->id)) : $currentReport = new ReportCard;
-                $currentReport->title = $report->title;
-                $currentReport->term_start = $report->term_start;
-                $currentReport->term_end = $report->term_end;
-                $currentReport->user_id = $user->id;
-                
-                $filePath = "belongs_to_rc_".strval($report->index);
-                if ($request->file($filePath)){
-                    $filenamewithextension = $request->file($filePath)->getClientOriginalName(); 	//get filename with extension
-                    $filename = pathinfo($filenamewithextension, PATHINFO_FILENAME); //get filename without extension
-                    $extension = $request->file($filePath)->getClientOriginalExtension(); //get file extension
-                    $filenametostore =  "Users/ReportCards/User_".$user->id."/".$filename.'_'.time().'.'.$extension;	//filename to store
-                    Storage::disk('s3')->put($filenametostore, fopen($request->file($filePath), 'r+'), 'public');	//Upload File to s3
-                    $report_url = "https://s3-ap-southeast-1.amazonaws.com/advoedu-testing/".$filenametostore;
-                    $currentReport->file = $report_url;
-                } else {
-                  $currentReport->file = $report->file;
+        if(isset($userParams->newReportCards)) {
+            $reportCards = $userParams->newReportCards;
+            foreach($reportCards as $report){
+                if($report->deleted == false){
+                    is_numeric($report->id) ? $currentReport = (ReportCard::find($report->id)) : $currentReport = new ReportCard;
+                    $currentReport->title = $report->title;
+                    $currentReport->term_start = $report->term_start;
+                    $currentReport->term_end = $report->term_end;
+                    $currentReport->user_id = $user->id;
+                    
+                    $filePath = "belongs_to_rc_".strval($report->index);
+                    if ($request->file($filePath)){
+                        $filenamewithextension = $request->file($filePath)->getClientOriginalName(); 	//get filename with extension
+                        $filename = pathinfo($filenamewithextension, PATHINFO_FILENAME); //get filename without extension
+                        $extension = $request->file($filePath)->getClientOriginalExtension(); //get file extension
+                        $filenametostore =  "Users/ReportCards/User_".$user->id."/".$filename.'_'.time().'.'.$extension;	//filename to store
+                        Storage::disk('s3')->put($filenametostore, fopen($request->file($filePath), 'r+'), 'public');	//Upload File to s3
+                        $report_url = "https://s3-ap-southeast-1.amazonaws.com/advoedu-testing/".$filenametostore;
+                        $currentReport->file = $report_url;
+                    } else {
+                      $currentReport->file = $report->file;
+                    }
+                    $currentReport->save();
                 }
-                $currentReport->save();
-            }
-            if ($report->deleted == true && is_numeric($report->id)){
-                $currentReport = ReportCard::find($report->id);
-                $this->awsService->removeUpload($currentReport, $currentReport->file, "Users/ReportCards/User_".$user->id."/");
-                $currentReport->delete();
+                if ($report->deleted == true && is_numeric($report->id)){
+                    $currentReport = ReportCard::find($report->id);
+                    $this->awsService->removeUpload($currentReport, $currentReport->file, "Users/ReportCards/User_".$user->id."/");
+                    $currentReport->delete();
+                }
             }
         }
+
+        // Update funding targets
+        dd($userParams);
+        if(isset($userParams->fundingTargets)){
+            foreach($userParams->fundingTargets as $ft){
+                if($ft->deleted == false){
+                    is_numeric($ft->id) ? $currentFt = (FundingTarget::find($ft->id)) : $currentFt = new FundingTarget;
+                    $currentFt->title = $ft->title;
+                    $currentFt->amount = $ft->amount;
+                    $currentFt->save();
+                }
+            }
+        };
         return view('admin.users.index');
     }
